added public announcements route
