+++ api/src/main/java/com/codeforcommunity/api/IAnnouncementsProcessor.java
@@ -1,21 +1,12 @@
 package com.codeforcommunity.api;
 
 import com.codeforcommunity.auth.JWTData;
-import com.codeforcommunity.dto.announcements.GetAnnouncementsRequest;
 import com.codeforcommunity.dto.announcements.GetAnnouncementsResponse;
 import com.codeforcommunity.dto.announcements.GetEventSpecificAnnouncementsRequest;
 import com.codeforcommunity.dto.announcements.PostAnnouncementRequest;
 import com.codeforcommunity.dto.announcements.PostAnnouncementResponse;
 
 public interface IAnnouncementsProcessor {
-
-  /**
-   * Gets all announcements.
-   *
-   * @param request DTO containing optional params, startDate, endDate, and count
-   * @return an announcements response DTO
-   */
-  GetAnnouncementsResponse getAnnouncements(GetAnnouncementsRequest request);
 
   /**
    * Creates a new announcement.
