+++ api/src/main/java/com/codeforcommunity/rest/ApiRouter.java
@@ -5,6 +5,7 @@ import com.codeforcommunity.api.IAuthProcessor;
 import com.codeforcommunity.api.ICheckoutProcessor;
 import com.codeforcommunity.api.IEventsProcessor;
 import com.codeforcommunity.api.IProtectedUserProcessor;
+import com.codeforcommunity.api.IPublicAnnouncementsProcessor;
 import com.codeforcommunity.api.IPublicEventsProcessor;
 import com.codeforcommunity.api.IRequestsProcessor;
 import com.codeforcommunity.auth.JWTAuthorizer;
@@ -15,6 +16,7 @@ import com.codeforcommunity.rest.subrouter.CommonRouter;
 import com.codeforcommunity.rest.subrouter.EventsRouter;
 import com.codeforcommunity.rest.subrouter.PfRequestRouter;
 import com.codeforcommunity.rest.subrouter.ProtectedUserRouter;
+import com.codeforcommunity.rest.subrouter.PublicAnnouncementsRouter;
 import com.codeforcommunity.rest.subrouter.PublicEventsRouter;
 import com.codeforcommunity.rest.subrouter.WebhooksRouter;
 import io.vertx.core.Vertx;
@@ -29,6 +31,7 @@ public class ApiRouter implements IRouter {
   private final EventsRouter eventsRouter;
   private final PublicEventsRouter publicEventsRouter;
   private final AnnouncementsRouter announcementsRouter;
+  private final PublicAnnouncementsRouter publicAnnouncementsRouter;
   private final CheckoutRouter checkoutRouter;
   private final WebhooksRouter webhooksRouter;
 
@@ -39,6 +42,7 @@ public class ApiRouter implements IRouter {
       IEventsProcessor eventsProcessor,
       IPublicEventsProcessor publicEventsProcessor,
       IAnnouncementsProcessor announcementEventsProcessor,
+      IPublicAnnouncementsProcessor publicAnnouncementsProcessor,
       ICheckoutProcessor checkoutProcessor,
       JWTAuthorizer jwtAuthorizer) {
 
@@ -49,6 +53,7 @@ public class ApiRouter implements IRouter {
     this.eventsRouter = new EventsRouter(eventsProcessor);
     this.publicEventsRouter = new PublicEventsRouter(publicEventsProcessor);
     this.announcementsRouter = new AnnouncementsRouter(announcementEventsProcessor);
+    this.publicAnnouncementsRouter = new PublicAnnouncementsRouter(publicAnnouncementsProcessor);
     this.checkoutRouter = new CheckoutRouter(checkoutProcessor);
     this.webhooksRouter = new WebhooksRouter(checkoutProcessor);
   }
@@ -60,6 +65,7 @@ public class ApiRouter implements IRouter {
     router.mountSubRouter("/user", authRouter.initializeRouter(vertx));
     router.mountSubRouter("/webhooks", webhooksRouter.initializeRouter(vertx));
     router.mountSubRouter("/events", publicEventsRouter.initializeRouter(vertx));
+    router.mountSubRouter("/announcements", publicAnnouncementsRouter.initializeRouter(vertx));
     router.mountSubRouter("/protected", defineProtectedRoutes(vertx));
 
     return router;
