+++ service/src/test/java/com/codeforcommunity/processor/AnnouncementsProcessorImplTest.java
@@ -1,11 +1,7 @@
 package com.codeforcommunity.processor;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.fail;
-
 import com.codeforcommunity.JooqMock;
 import com.codeforcommunity.auth.JWTData;
-import com.codeforcommunity.dto.announcements.GetAnnouncementsRequest;
 import com.codeforcommunity.dto.announcements.GetAnnouncementsResponse;
 import com.codeforcommunity.dto.announcements.GetEventSpecificAnnouncementsRequest;
 import com.codeforcommunity.dto.announcements.PostAnnouncementRequest;
@@ -14,14 +10,18 @@ import com.codeforcommunity.enums.PrivilegeLevel;
 import com.codeforcommunity.exceptions.AdminOnlyRouteException;
 import com.codeforcommunity.exceptions.MalformedParameterException;
 import com.codeforcommunity.requester.Emailer;
-import java.sql.Timestamp;
-import java.util.ArrayList;
-import java.util.List;
 import org.jooq.generated.Tables;
 import org.jooq.generated.tables.records.AnnouncementsRecord;
 import org.jooq.generated.tables.records.EventsRecord;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import java.sql.Timestamp;
+import java.util.ArrayList;
+import java.util.List;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.fail;
 
 // Contains tests for AnnouncementsProcessorImpl.java in main
 public class AnnouncementsProcessorImplTest {
@@ -44,78 +44,6 @@ public class AnnouncementsProcessorImplTest {
     this.myAnnouncementsProcessorImpl =
         new AnnouncementsProcessorImpl(
             myJooqMock.getContext(), new Emailer(myJooqMock.getContext()));
-  }
-
-  // test getting announcements with range covering no events
-  @Test
-  public void testGetAnnouncements1() {
-    // craft the get request
-    GetAnnouncementsRequest req =
-        new GetAnnouncementsRequest(new Timestamp(0), new Timestamp(1000), 2);
-
-    myJooqMock.addEmptyReturn("SELECT");
-    GetAnnouncementsResponse res = myAnnouncementsProcessorImpl.getAnnouncements(req);
-
-    assertEquals(res.getTotalCount(), 0);
-  }
-
-  // test getting announcements with range covering all events
-  @Test
-  public void testGetAnnouncements2() {
-    // craft the get request
-    GetAnnouncementsRequest req =
-        new GetAnnouncementsRequest(
-            new Timestamp(START_TIMESTAMP_TEST2 - 1000),
-            new Timestamp(END_TIMESTAMP_TEST + 1000),
-            2);
-
-    // mock the announcements inside the DB
-    AnnouncementsRecord announcement1 = myJooqMock.getContext().newRecord(Tables.ANNOUNCEMENTS);
-    announcement1.setId(0);
-    announcement1.setTitle("the first announcement title");
-    announcement1.setCreated(new Timestamp(START_TIMESTAMP_TEST));
-    announcement1.setDescription("the first announcement description");
-
-    AnnouncementsRecord announcement2 = myJooqMock.getContext().newRecord(Tables.ANNOUNCEMENTS);
-    announcement2.setId(1);
-    announcement2.setTitle("the second announcement title");
-    announcement2.setCreated(new Timestamp(START_TIMESTAMP_TEST2));
-    announcement2.setDescription("the second announcement description");
-
-    List<AnnouncementsRecord> announcements = new ArrayList<>();
-    announcements.add(announcement1);
-    announcements.add(announcement2);
-    myJooqMock.addReturn("SELECT", announcements);
-
-    GetAnnouncementsResponse res = myAnnouncementsProcessorImpl.getAnnouncements(req);
-
-    assertEquals(res.getTotalCount(), 2);
-    assertEquals(res.getAnnouncements().get(0).getId(), 0);
-    assertEquals(res.getAnnouncements().get(1).getId(), 1);
-  }
-
-  // test getting announcements with range covering some events
-  @Test
-  public void testGetAnnouncements3() {
-    // craft the get request
-    GetAnnouncementsRequest req =
-        new GetAnnouncementsRequest(
-            new Timestamp(START_TIMESTAMP_TEST - 1000),
-            new Timestamp(END_TIMESTAMP_TEST + 1000),
-            2);
-
-    // mock the announcements inside the DB
-    AnnouncementsRecord announcement1 = myJooqMock.getContext().newRecord(Tables.ANNOUNCEMENTS);
-    announcement1.setId(0);
-    announcement1.setTitle("the first announcement title");
-    announcement1.setCreated(new Timestamp(START_TIMESTAMP_TEST));
-    announcement1.setDescription("the first announcement description");
-    myJooqMock.addReturn("SELECT", announcement1);
-
-    GetAnnouncementsResponse res = myAnnouncementsProcessorImpl.getAnnouncements(req);
-
-    assertEquals(res.getTotalCount(), 1);
-    assertEquals(res.getAnnouncements().get(0).getId(), 0);
   }
 
   // test posting an announcement without admin privileges
